Player ranking now also for past years
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@11248 8aca7d54-2c30-db11-9de9-000461428c89

+++ src/simutrans/utils/cbuffer.h
@@ -16,12 +16,12 @@ class cbuffer_t
 {
 private:
 
-	unsigned int capacity;
+	size_t capacity;
 
 	/**
 	* Number of characters without(!) trailing '\0'
 	*/
-	unsigned int size;
+	size_t size;
 
 	char *buf;
